requirements.txt
Django==1.6.4
Pillow
